+++ test/acceptance/authz/swagger_helper.go
@@ -9,7 +9,7 @@
 //  CONTACT: hello@weaviate.io
 //
 
-package test
+package authz
 
 import (
 	"encoding/json"
